requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.5.4
Django==5.0.7
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
drf-access-policy==1.5.0
drf-flex-fields==1.0.2
elastic-transport==8.13.1
elasticsearch==8.14.0
factory-boy==3.3.0
Faker==26.0.0
h11==0.14.0
iniconfig==2.0.0
kombu==5.3.7
packaging==24.1
pika==1.3.2
pillow==10.3.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg==3.2.1
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
redis==5.0.7
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
